refactor: entity のテストが自 Slice の barrel を参照するのをやめる (FRESTYLE-165)
テストが @/entities/<自分> を参照すると Slice 内の全ファイルが読み込まれ、
カバレッジの分母に未テストのファイルまで入ってしまう。相対パスで必要な
モジュールだけを参照する。
