+++ frontend/src/entities/learning-report/ui/__tests__/ReportCard.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import ReportCard from '../ReportCard';
-import type { LearningReport } from '@/entities/learning-report';
+import type { LearningReport } from '../../model/types';
 
 function makeReport(overrides: Partial<LearningReport> = {}): LearningReport {
   return {
