+++ frontend/src/entities/notification/ui/__tests__/NotificationItem.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
 import NotificationItem from '../NotificationItem';
-import type { Notification } from '@/entities/notification';
+import type { Notification } from '../../model/types';
 
 const baseNotification: Notification = {
   id: 1,
